requirements.txt
Flask==3.1.3
flask-cors
google-generativeai==0.8.3
python-dotenv==1.0.1
playwright==1.49.0